encriptio functions to own .py file
