+++ tests/modules/duckdb_utils/test_add_primary_key.py
