+++ tests/modules/duckdb_utils/test_check_if_primary_key_exists.py
