+++ tests/modules/duckdb_utils/test_get_db_schema_tbl_from_table_name.py
